test(validation): split the wiring tests at the file-size ceiling
The scope file crossed 500 lines when the round-5 regressions landed;
TestPrePrWiring moves to its own file along the class seam, matching
the session's extract-at-the-seam lesson.

Refs #5337
